Preserve auth connection ids in graph serialization
